# docs/swagger.yaml
basePath: /api
definitions:
...
  web_backend_internal_app_serializer.DepartmentApplicationJSON:
    properties:
      created_at:
        type: string
      creator_login:
        type: string
      department_application_id:
        type: integer
      finish_date:
        type: string
      forming_date:
        type: string
      incomplete_items_count:
        type: integer
      moderator_login:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  web_backend_internal_app_serializer.DepartmentJSON:
    properties:
      department_id:
        type: integer
      description:
        type: string
      employee_count:
        type: integer
      head:
        type: string
      is_deleted:
        type: boolean
      photo_url:
        type: string
      reports_to:
        type: string
      short_description:
        type: string
      title:
        type: string
      video:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://localhost:8080
  description: API для управления заявками отделов
  license:
    name: MIT
  title: Department Application API
  version: "1.0"
paths:
...
  /dep_app_dep/add/{department_id}:
    post:
      description: Добавляет отдел в заявку-черновик пользователя
      parameters:
      - description: ID отдела
        in: path
        name: department_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Заявка с добавленным отделом
          schema:
            $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentApplicationJSON'
        "201":
          description: Создана новая заявка
          schema:
            $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentApplicationJSON'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Отдел не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Отдел уже в заявке
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить отдел в заявку
      tags:
      - department_application_departments
  /department/{id}:
    get:
      description: Возвращает информацию об отделе по идентификатору
      parameters:
      - description: ID отдела
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные отдела
          schema:
            $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentJSON'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Отдел не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить отдел по ID
      tags:
      - departments
  /department/create-department:
    post:
      consumes:
      - application/json
      description: Создает новый отдел
      parameters:
      - description: Данные нового отдела
        in: body
        name: department
        required: true
        schema:
          $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Созданный отдел
          schema:
            $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentJSON'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать отдел
      tags:
      - departments
...
  /departments:
    get:
      description: Возвращает все отделы или фильтрует по названию
      parameters:
      - description: Название отдела для поиска
        in: query
        name: Title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список отделов
          schema:
            items:
              $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentJSON'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список отделов
      tags:
      - departments
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"